move the definition of _LOC to general
